#18 [Refactor] StoreId Long으로 전환

+++ src/main/java/com/mpnp/baechelin/api/model/PublicApiV1Form.java
@@ -46,7 +46,7 @@ public class PublicApiV1Form {
     public static class Row {
         @JsonProperty("SEQ")
         int SEQ;
-        Integer storeId;
+        Long storeId;
         BigDecimal latitude;
         BigDecimal longitude;
         String category;
